deleted:    web/controllers/admin_controller.ex 	new file:   web/controllers/comment_controller.ex 	modified:   web/controllers/post_controller.ex 	modified:   web/helpers/auth.ex 	modified:   web/router.ex 	deleted:    web/templates/admin/index.html.eex 	deleted:    web/templates/admin/new.html.eex 	deleted:    web/templates/admin/show.html.eex 	modified:   web/templates/layout/app.html.eex 	deleted:    web/templates/page/index.html.eex 	modified:   web/templates/post/index.html.eex 	modified:   web/templates/post/show.html.eex
