--- web/templates/admin/index.html.eex
@@ -1,19 +1,0 @@
-<div>
-  <h2># Admin Page</h2>
-  <div>
-    <p><b>## Posts:</b></p>
-  </div>
-  <div>
-    <ul>
-      <%= for post <- @posts do %>
-          <li>[<%= post.updated_at %>] <b><%= post.title %>:</b> <%= post.body %>
-            <br>
-            <%#= link "Delete", to: admin_path(@conn, :delete, post), method: :delete %>
-            <%#= link "Edit",   to: admin_path(@conn, :update, post), method: :get %>
-          </li>
-      <% end %>
-    </ul>
-  </div>
-  <br>
-  <div><a href="/admin/new/">Create</a></div>
-</div>
